requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.9.0
fastapi==0.110.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pytest==8.1.1
python-dateutil==2.9.0.post0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.11.0
uvicorn==0.29.0
